# specs/core-gateway.openapi.yaml
openapi: 3.1.0
info:
  title: StarlingAI Gateway API
  version: 0.7.0
  summary: HTTP + WebSocket surface for the StarlingAI agent-swarm gateway.
  description: |
    REST API for the `@starlingai/core` gateway. Companion WebSocket RPC is at
    `ws://${host}:${port}/ws?token=<jwt>` and follows JSON-RPC 2.0 (see
    `docs/api.md` for the RPC method catalog).

    Authentication: all `/api/*` paths require a bearer JWT unless noted.
    Trigger endpoints also accept a webhook key via `?key=` or a header — but the
    header name differs per endpoint: `POST /api/scenes/{name}/run` reads
    `X-Scene-Key`, while `POST /api/jobs/{name}/run` reads `X-Job-Key`.

    Schemas below are intentionally loose — the runtime validates with Zod
    schemas in `packages/core/src/config/` and `packages/core/src/gateway/`.
    Refer to those schemas for exhaustive field definitions.
  license:
    name: CC BY-NC 4.0
    url: https://creativecommons.org/licenses/by-nc/4.0/
servers:
  - url: http://localhost:8765
    description: Local default (gateway.port)
  - url: https://{host}
    description: Reverse-proxied deployment
    variables:
      host:
        default: ai.internal.example
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    sceneKey:
      type: apiKey
      in: header
      name: X-Scene-Key
    jobKey:
      type: apiKey
      in: header
      name: X-Job-Key
  schemas:
...
    ChannelConfig:
      type: object
      additionalProperties: true
...
security:
  - bearerAuth: []
paths:
...
  /api/channels/{type}:
    get:
      tags: [Channels]
      summary: Read a channel's runtime state + config
      parameters: [ { name: type, in: path, required: true, schema: { type: string, enum: [slack, telegram, discord, whatsapp, email, webchat] } } ]
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ChannelConfig' } } } } }
    put:
      tags: [Channels]
      summary: Upsert a channel config
      parameters: [ { name: type, in: path, required: true, schema: { type: string } } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ChannelConfig' } } } }
      responses: { '200': { description: Saved } }
    delete:
      tags: [Channels]
      summary: Disable a channel
      parameters: [ { name: type, in: path, required: true, schema: { type: string } } ]
      responses: { '200': { description: 'Deleted (200 with a JSON body, not 204)', content: { application/json: { schema: { type: object, properties: { ok: { type: boolean } } } } } } }
...